refactor component into its own file, move shared deps to utils.js
